+++ test/document_stores/test_opensearch.py
@@ -19,15 +19,15 @@ from haystack.document_stores.opensearch import (
 from haystack.schema import Document, Label, Answer
 from haystack.errors import DocumentStoreError
 
-# Being all the tests in this module, ideally we wouldn't need a marker here,
-# but this is to allow this test suite to be skipped when running (e.g.)
-# `pytest test/document_stores --document-store-type=faiss`
-class TestOpenSearchDocumentStore:
+from .test_base import DocumentStoreTest
+
+
+class TestOpenSearchDocumentStore(DocumentStoreTest):
 
     # Constants
 
     query_emb = np.random.random_sample(size=(2, 2))
-    index_name = "myindex"
+    index_name = __name__
 
     # Fixtures
 
@@ -36,11 +36,10 @@ class TestOpenSearchDocumentStore:
         """
         This fixture provides a working document store and takes care of removing the indices when done
         """
-        index_name = __name__
-        labels_index_name = f"{index_name}_labels"
-        ds = OpenSearchDocumentStore(index=index_name, label_index=labels_index_name, port=9201, create_index=True)
+        labels_index_name = f"{self.index_name}_labels"
+        ds = OpenSearchDocumentStore(index=self.index_name, label_index=labels_index_name, port=9201, create_index=True)
         yield ds
-        ds.delete_index(index_name)
+        ds.delete_index(self.index_name)
         ds.delete_index(labels_index_name)
 
     @pytest.fixture
@@ -83,35 +82,6 @@ class TestOpenSearchDocumentStore:
         }
 
     @pytest.fixture
-    def documents(self):
-        documents = []
-        for i in range(3):
-            documents.append(
-                Document(
-                    content=f"A Foo Document {i}",
-                    meta={"name": f"name_{i}", "year": "2020", "month": "01"},
-                    embedding=np.random.rand(768).astype(np.float32),
-                )
-            )
-
-            documents.append(
-                Document(
-                    content=f"A Bar Document {i}",
-                    meta={"name": f"name_{i}", "year": "2021", "month": "02"},
-                    embedding=np.random.rand(768).astype(np.float32),
-                )
-            )
-
-            documents.append(
-                Document(
-                    content=f"Document {i} without embeddings",
-                    meta={"name": f"name_{i}", "no_embedding": True, "month": "03"},
-                )
-            )
-
-        return documents
-
-    @pytest.fixture
     def index(self):
         return {
             "aliases": {},
@@ -143,23 +113,6 @@ class TestOpenSearchDocumentStore:
             },
         }
 
-    @pytest.fixture
-    def labels(self, documents):
-        labels = []
-        for i, d in enumerate(documents):
-            labels.append(
-                Label(
-                    query="query",
-                    document=d,
-                    is_correct_document=True,
-                    is_correct_answer=False,
-                    # create a mix set of labels
-                    origin="user-feedback" if i % 2 else "gold-label",
-                    answer=None if not i else Answer(f"the answer is {i}"),
-                )
-            )
-        return labels
-
     # Integration tests
 
     @pytest.mark.integration
@@ -169,20 +122,6 @@ class TestOpenSearchDocumentStore:
     @pytest.mark.integration
     def test___init___faiss(self):
         OpenSearchDocumentStore(index="faiss_index", port=9201, create_index=True, knn_engine="faiss")
-
-    @pytest.mark.integration
-    def test_write_documents(self, ds, documents):
-        ds.write_documents(documents)
-        docs = ds.get_all_documents()
-        assert len(docs) == len(documents)
-        for i, doc in enumerate(docs):
-            expected = documents[i]
-            assert doc.id == expected.id
-
-    @pytest.mark.integration
-    def test_write_labels(self, ds, labels):
-        ds.write_labels(labels)
-        assert ds.get_all_labels() == labels
 
     @pytest.mark.integration
     def test_recreate_index(self, ds, documents, labels):
